@@ -1,33 +1,35 @@
 {
   "devDependencies": {
-    "@fortawesome/fontawesome-free": "^5.10.1",
+    "@fortawesome/fontawesome-free": "^5.12.0",
+    "@sentry/cli": "^1.49.0",
+    "@sentry/electron": "^1.0.0",
     "@types/electron-config": "^3.2.2",
     "@types/electron-debug": "^2.1.0",
-    "@types/fs-promise": "1.0.3",
     "@types/js-yaml": "^3.12.1",
-    "@types/node": "12.6.2",
-    "@types/webpack-env": "1.13.9",
-    "@typescript-eslint/eslint-plugin": "^1.13.0",
-    "@typescript-eslint/parser": "^1.13.0",
-    "app-builder-lib": "^21.0.8",
+    "@types/node": "12.7.12",
+    "@types/webpack-env": "1.14.1",
+    "@typescript-eslint/eslint-plugin": "^2.13.0",
+    "@typescript-eslint/parser": "^2.14.0",
     "apply-loader": "2.0.0",
     "awesome-typescript-loader": "^5.0.0",
-    "core-js": "^3.2.0",
-    "cross-env": "5.2.0",
-    "css-loader": "3.2.0",
-    "electron": "^5.0.7",
-    "electron-builder": "^21.0.8",
-    "electron-installer-snap": "^4.0.0",
+    "core-js": "^3.6.1",
+    "cross-env": "6.0.3",
+    "css-loader": "3.4.0",
+    "electron": "^7.1.7",
+    "electron-builder": "22.1.0",
+    "electron-download": "^4.1.1",
+    "electron-installer-snap": "^4.1.0",
     "electron-notarize": "^0.1.1",
     "electron-rebuild": "^1.8.5",
-    "eslint": "^5.16.0",
-    "file-loader": "^4.1.0",
-    "graceful-fs": "^4.2.1",
+    "eslint": "^6.8.0",
+    "eslint-plugin-import": "^2.19.1",
+    "file-loader": "^5.0.2",
+    "graceful-fs": "^4.2.2",
     "html-loader": "0.5.5",
     "json-loader": "0.5.7",
-    "node-abi": "^2.11.0",
-    "node-gyp": "^5.0.3",
-    "node-sass": "^4.12.0",
+    "node-abi": "^2.12.0",
+    "node-gyp": "^6.0.1",
+    "node-sass": "^4.13.0",
     "npmlog": "4.1.2",
     "npx": "^10.2.0",
     "pug": "^2.0.4",
@@ -35,22 +37,21 @@
     "pug-lint": "^2.6.0",
     "pug-loader": "^2.4.0",
     "pug-static-loader": "2.0.0",
-    "raven-js": "3.27.2",
-    "raw-loader": "3.1.0",
-    "sass-loader": "^7.2.0",
+    "raw-loader": "4.0.0",
+    "sass-loader": "^8.0.0",
     "shelljs": "0.8.3",
-    "source-code-pro": "^2.30.1",
-    "source-sans-pro": "2.45.0",
-    "style-loader": "^1.0.0",
+    "source-code-pro": "^2.30.2",
+    "source-sans-pro": "3.6.0",
+    "style-loader": "^1.1.2",
     "svg-inline-loader": "^0.8.0",
-    "to-string-loader": "1.1.5",
+    "to-string-loader": "1.1.6",
     "tslib": "^1.10.0",
-    "typedoc": "^0.15.0",
-    "typescript": "^3.5.3",
-    "url-loader": "^2.1.0",
-    "val-loader": "1.1.1",
-    "webpack": "^4.39.1",
-    "webpack-cli": "^3.3.6",
+    "typedoc": "^0.15.6",
+    "typescript": "^3.7.4",
+    "url-loader": "^3.0.0",
+    "val-loader": "2.1.0",
+    "webpack": "^5.0.0-beta.11",
+    "webpack-cli": "^3.3.10",
     "yaml-loader": "0.5.0"
   },
   "resolutions": {
@@ -80,11 +81,6 @@
       "installerIcon": "./build/windows/icon.ico"
     },
     "publish": [
-      {
-        "provider": "bintray",
-        "token": "d993c4faa708a4cba84fa3a8e822457e7298d75c",
-        "component": "main"
-      },
       {
         "provider": "github"
       }
@@ -140,8 +136,5 @@
     "lint": "eslint --ext ts */src",
     "postinstall": "node ./scripts/install-deps.js"
   },
-  "repository": "eugeny/terminus",
-  "dependencies": {
-    "eslint-plugin-import": "^2.18.2"
-  }
+  "repository": "eugeny/terminus"
 }
