@@ -2,11 +2,12 @@
 const builder = require('electron-builder').build
 const vars = require('./vars')
 
-const isTag = (process.env.BUILD_SOURCEBRANCH || '').startsWith('refs/tags/')
+const isTag = (process.env.GITHUB_REF || '').startsWith('refs/tags/')
+const isCI = !!process.env.GITHUB_REF
 
 builder({
     dir: true,
-    linux: ['snap', 'deb', 'rpm', 'tar.gz'],
+    linux: ['deb', 'tar.gz', 'rpm'],
     config: {
         extraMetadata: {
             version: vars.version,
