@@ -56,10 +56,6 @@ export class ElectronService {
         browserWindow: Electron.BrowserWindow,
         options: Electron.MessageBoxOptions
     ): Promise<MessageBoxResponse> {
-        return new Promise(resolve => {
-            this.dialog.showMessageBox(browserWindow, options, (response, checkboxChecked) => {
-                resolve({ response, checkboxChecked })
-            })
-        })
+        return this.dialog.showMessageBox(browserWindow, options)
     }
 }
