@@ -6,7 +6,7 @@ import { SettingsTabComponent } from './components/settingsTab.component'
 /** @hidden */
 @Injectable()
 export class RecoveryProvider extends TabRecoveryProvider {
-    async recover (recoveryToken: any): Promise<RecoveredTab> {
+    async recover (recoveryToken: any): Promise<RecoveredTab|null> {
         if (recoveryToken && recoveryToken.type === 'app:settings') {
             return { type: SettingsTabComponent }
         }
