@@ -8,7 +8,7 @@ import { TerminalColorScheme } from './api/interfaces'
 @Injectable()
 export class HyperColorSchemes extends TerminalColorSchemeProvider {
     async getSchemes (): Promise<TerminalColorScheme[]> {
-        const pluginsPath = path.join(process.env.HOME, '.hyper_plugins', 'node_modules')
+        const pluginsPath = path.join(process.env.HOME as string, '.hyper_plugins', 'node_modules')
         if (!await fs.exists(pluginsPath)) {
             return []
         }
