@@ -6,6 +6,7 @@ export interface SearchOptions {
     regex?: boolean
     wholeWord?: boolean
     caseSensitive?: boolean
+    incremental?: true
 }
 
 /**
@@ -23,7 +24,7 @@ export abstract class Frontend {
     protected mouseEvent = new Subject<MouseEvent>()
     protected bell = new Subject<void>()
     protected contentUpdated = new Subject<void>()
-    protected input = new Subject<string>()
+    protected input = new Subject<Buffer>()
     protected resize = new ReplaySubject<ResizeEvent>(1)
     protected dragOver = new Subject<DragEvent>()
     protected drop = new Subject<DragEvent>()
@@ -34,7 +35,7 @@ export abstract class Frontend {
     get mouseEvent$ (): Observable<MouseEvent> { return this.mouseEvent }
     get bell$ (): Observable<void> { return this.bell }
     get contentUpdated$ (): Observable<void> { return this.contentUpdated }
-    get input$ (): Observable<string> { return this.input }
+    get input$ (): Observable<Buffer> { return this.input }
     get resize$ (): Observable<ResizeEvent> { return this.resize }
     get dragOver$ (): Observable<DragEvent> { return this.dragOver }
     get drop$ (): Observable<DragEvent> { return this.drop }
